Split `for_loops` tests
